ab_example_contract_wallet::tx_handler_ffi::execute

Function tx_handler_execute

#[unsafe(no_mangle)]
pub unsafe extern "C" fn tx_handler_execute(
    args: NonNull<InternalArgs<'_>>,
) -> ExitCode

FFI interface into a method, called by the host.

NOTE: Calling this function directly shouldn’t be necessary except maybe in contract’s own tests.

Safety

Caller must ensure the provided pointer corresponds to expected ABI.